ai_sidecar/requirements.txt
# AI Sidecar Core Dependencies
# 
# Install with: pip install -r requirements.txt
# For development: pip install -r requirements.txt -e ".[dev]"

# Core Framework
pyzmq>=25.1.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
structlog>=23.2.0
aiofiles>=23.2.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Memory and Learning (Phase 8)
redis>=5.0.0
openai>=1.68.0
anthropic>=0.30.0
httpx>=0.27.0

# ML and Data Processing
numpy>=1.24.0
scikit-learn>=1.3.0

# Azure OpenAI (Optional)
azure-identity>=1.14.0
azure-ai-openai>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0